Add clicking on stations for route

+++ src/main/java/seng202/team6/controller/DataToolBarController.java
@@ -51,7 +51,7 @@ public class DataToolBarController implements ScreenController {
                     + "%') AND ";
         }
         if (distanceSliderOfFilter.getValue() != 0) {
-            float[] latlng = controller.getMapController().getLatLng();
+            double[] latlng = controller.getMapController().getLatLng();
             if (latlng[0] == 0) {
                 AlertMessage.createMessage("Current Location has not been selected.",
                         "Please select a location on the map.");
